requirements.txt
allure-pytest==2.12.0
allure-python-commons==2.12.0
pytest==7.2.1
requests==2.28.2
selenium==4.27.1